requirements.txt
Flask==3.1.*
flask-cors>=5.0
gunicorn>=22.0
litellm>=1.40
pdfplumber>=0.11
python-dotenv>=1.0
